+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased][]
 
+### Changed
+
+- Use `useEffect` over `useLayoutEffect` for deferred performance reasons
+
 ## [0.2.2][] - 2019-02-02
 
 ### Fixed